\name{mhrcmxIntDep-package}
\alias{mhrcmxIntDep-package}
\alias{mhrcmxIntDep}
\docType{package}
\title{\packageTitle{mhrcmxIntDep}}
\description{\packageDescription{mhrcmxIntDep}}
\section{Package Content}{\packageIndices{mhrcmxIntDep}}
\author{\packageAuthor{mhrcmxIntDep}}
\section{Maintainer}{\packageMaintainer{mhrcmxIntDep}}
\keyword{package}
